requirements.txt
# EFA-ECG requirements
# Python 3.10

# Deep learning
torch>=2.2.0
transformers>=4.44.0

# ECG signal processing
wfdb>=4.1.0
neurokit2>=0.2.0

# NLP
spacy>=3.7.0

# Data & scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Google Gemini API
google-generativeai>=0.5.0

# Utilities
tqdm>=4.65.0
Pillow>=10.0.0
